All 18 CMake targets build clean: 0 compile errors, 0 link errors
MILESTONE: Full project compiles and links successfully.

Core targets built:
- MovieMaker.exe (thin launcher, delay-loads MovieMakerCore.dll)
- MovieMakerCore.dll (main codebase, 1MB)
- WLXPhotoBase.dll (foundation library)
- MovieMakerLang.dll
- WLXPipeline.dll, WLXVideoTrim.dll

All 12 external stub DLLs also build:
- WLXMovieLibrary, WLXPipetran, WLXSlideshow, WLXTranscode
- MovieMakerPreviewClient, WLMFDS, WLMFReadWrite
- WLXMediaPublishSubscribe, WLXPhotoCinematic, WLXMP4Parser
- WLXFaceRecognition, WLXCodecHost

Key fixes in this session:
- Added stub implementations: ThumbnailCache, UndoManager, TimelineController,
  ProjectManager, ExportController, ImportController, PlaybackController,
  ValueVariantRef, d3dx11compat, dxva2stubs
- Fixed X3DNodes.cpp: MF* member names m_value->m_values, inline SLERP
- Fixed MovieMakerCore CMakeLists: WMMR_DLL_EXPORTS, all new .cpp files,
  WLXPhotoBase link, removed uxctl.lib, d3dcompiler_46->d3dcompiler
- Fixed dllmain.cpp: extern C on MovieCore_Initialize/Shutdown
- Fixed exports.h: MovieMakerMain signature (int, wchar_t**)
- Fixed MeshResourceDX.cpp: Release() for Grid/Shatter/PageCurl/Wipe resources
- Fixed pch.h: IDXVA2 stub interface, DXVA2 function declarations
- Added WLXPhotoBase_Init export to WLXPhotoBase.dll
- Fixed all 12 external stub DLL errors:
  - D3D9 headers, std::memory, std::algorithm, std::set includes
  - Moved forward-declared structs into correct namespaces (6 targets)
  - Fixed DEFINE_STUB_TRANSITION macro token paste
  - Fixed GDI+ const qualification, MFAttributes type, removed mferror.lib
  - Replaced INTERNET_OPTION_ENABLE_FEATURE and MF_ENABLE_HARDWARE_TRANSFORMS
  - Fixed duplicate DllMain, removed d3dx9 from all targets
  - Fixed CAtlArray, CComObject, std::min type mismatch issues

+++ src/MovieMaker/main.cpp
@@ -15,7 +15,7 @@ extern "C" int __cdecl MovieMakerMain(int argc, wchar_t** argv);
 // The original binary registers a __except handler around the
 // LoadLibrary/GetProcAddress region that catches C++ exceptions
 // with subcodes 0x19930520-22 and 0x01994000.
-static DWORD WINAPI VexHandler(PEXCEPTION_POINTERS ep)
+static LONG WINAPI VexHandler(PEXCEPTION_POINTERS ep)
 {
     if (ep->ExceptionRecord->ExceptionCode == 0xe06d7363)
     {
